consolidate original_data_test
